Introduce DiscreteField class.
Previously Element.gbasis was returning either 2-tuple or 3-tuple depending on
the type of element and its order-attribute. This return value was unpacked
and fed to the (bi)linear forms written by the user.

For simplifying the return value and enabling experiments with more generic form
definitions, we introduced a NamedTuple with optional attributes f, df, ddf (and
possibly more added later). This NamedTuple is called DiscreteField and
represents a global discrete finite element function (and its derivatives)
evaluated at the quadrature points.

Small backwards-incompatibility was introduced by the respective changes in
Basis.interpolate which now returns DiscreteField object (a NamedTuple with 3
values, i.e. 3-tuple) instead of 2-tuple and some code might depend on that
behavior. In the test suite there was only two small changes due to this.